v2.0-40

+++ drivers/isdn/icn/icn.c
@@ -202,6 +202,7 @@
  *
  */
 
+#include <linux/config.h>
 #include "icn.h"
 
 /*
@@ -772,7 +773,7 @@ icn_polldchan(unsigned long data)
 	int left;
 	u_char c;
 	int ch;
-	int flags;
+	unsigned long flags;
 	int i;
 	u_char *p;
 	isdn_ctrl cmd;
